fix thread-death LoadTestRunner
Previously, an ArrayBlockingQueue is used to prevent LoadTestRunner from
running out of memory.
However, when thread pool is fully loaded and the queue is full,
ThreadPoolExecutor rejects new tasks and call its
RejectedExecutionHandler.
By default, the handler throws an exception to the thread trying to
submit the new task.
Consequently, the load tester dies and the test hangs forever.

This commit fixes this by using DiscardPolicy, which silently
drops all rejected tasks.