add go publisher
This PR also fixes a race in AdapterTask.
Previously it issues a StartRequest to the worker process without
waiting for the RPC to complete.
Consequently, the load test might happen before the worker process
could process the StartRequest.
This is fixed by making AdapterTask wait for the request to return.