Add a queue to the batcher
The wait() function was showing up as the major limiting factor in profiles, so a queue was added to the batcher so that wait does not need to be called if values are immediately available.

Also modify batcher to enforce the 1000 message hard limit on publish requests while also enforcing a soft user-provided byte limit.